requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.28.0
numpy>=1.23.0
openai>=1.0.0

# Vector search
faiss-cpu>=1.7.4

# Progress bars
tqdm>=4.65.0

# Testing
pytest>=7.0.0
pytest-mock>=3.12.0

# Local model support (optional - only needed if USE_LOCAL_MODEL=true)
# Ollama is run via Docker or native installation
# No additional Python packages required for Ollama API client
